clarity rewrites
Proposed rewrites for clarity.  Does not fix the (two bullets I don't understand)[https://github.com/w3c/Guide/pull/112#issuecomment-769158287].  

"revisit" was not in a bullet item, so I moved it to the top.  If you want it in the last list, move it back, with the proper <LI> element?