clarity rewrites of timezone branch (#113)
* clarity rewrites

Proposed rewrites for clarity.  Does not fix the (two bullets I don't understand)[https://github.com/w3c/Guide/pull/112#issuecomment-769158287].  

"revisit" was not in a bullet item, so I moved it to the top.  If you want it in the last list, move it back, with the proper <LI> element?

* minor fixes to the edits

* further refinement

Took out the "revisit" item in the first list, since I think it's redundant with the guidance elsewhere.  Other minor changes.