Guidelines for timezones (#112)
* Guidelines for timezones

* small typo fixes

* clarity rewrites of timezone branch (#113)

* clarity rewrites

Proposed rewrites for clarity.  Does not fix the (two bullets I don't understand)[https://github.com/w3c/Guide/pull/112#issuecomment-769158287].  

"revisit" was not in a bullet item, so I moved it to the top.  If you want it in the last list, move it back, with the proper <LI> element?

* minor fixes to the edits

* further refinement

Took out the "revisit" item in the first list, since I think it's redundant with the guidance elsewhere.  Other minor changes.

* Update meetings/organize.html

Co-authored-by: Nigel Megitt <nigel.megitt@bbc.co.uk>

* Updated the tooling section

* Added dbaron suggestion for when to re-evaluate the time(s)

* From @JaninaSajka: consider holidays

* Majority of participants and balancing act

* Editorial pass (#119)

... including PLH's full signature with link back to GH

* Wording improvements from @samuelweiler

Co-authored-by: Ralph Swick <swick@w3.org>
Co-authored-by: Samuel Weiler <weiler@w3.org>
Co-authored-by: Nigel Megitt <nigel.megitt@bbc.co.uk>
Co-authored-by: Coralie Mercier <coralie@w3.org>