added "operations in travel restrictions" to section "guidelines and policies" (#127)
The W3C team travel policy lives as an archived mail to the W3C team and therefore is hard to link in /Guide or anywhere. But since we developed "Continuity of Operations under Travel Restrictions" which talks about travels (including w3c staff’s) and meetings during the pandemic, I see this as a good addition to the "guidelines and policies" section of the guidebook.

It also addresses one aspect of the [Advisory Board's issue-8](https://github.com/w3c/AB-memberonly/issues/8) that @chaals raised late 2019 (and that I became aware of today).