Editorial and QA on adv-notice.html (#219)
QA and updates (including to ensure matching "excerpted" texts from the
process and from the Tips document).
Also, I removed the excerpt from the Process at the bottom section on
"background", so that the document is shorter and we have one fewer
place to update if the Process text changes. The requirement in question
is linked from the top of the doc.