[coc-incident-resolution-ombuds] Adding revised text and new process with tEQuitable (#354)
This PR covers the revision of the document based on the addition of the
new ombuds process with tEQitable, as well as some editorial
suggestions.

Instead of merging PR #337 to an older version of the document, I have
adopted the changes to the introduction in this PR. We have not adopted
the change with the "in exceptional circumstances..." text as that line
is no longer needed in the document as a whole.

---------

Co-authored-by: Tzviya <trsiegman@gmail.com>
Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>