Used advisory terminology in guideline document (#355)
This document is a guideline. No matter how good its advice is, it is
only advice. Except when conveying requirements established elsewhere
authoritatively, guidelines should refrain from using phrasing like "is
required", because guidelines cannot actually require anything.

In this PR, I suggest changing one such instance from "is required" to
"is advised". No disagreement about the advice given, but as it isn't an
actual formal requirement, it should not be presented as such.