Add considerations on the final stage of documents (#366)
This creates a new page in the Guidebook that reviews considerations for
Working Groups pondering whether to publish the latest state of their
work as Candidate Recommendation (with Snapshots) or as Recommendations.

Four dimensions are described:
- targeted audience
- wide review
- stability/availability signals
- revisions/maintenance.

The revisions/maintenance section could perhaps be expanded and moved to
a dedicated page over time.

Note: This is intended to supersede #151.

---------

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>
Co-authored-by: Nigel Megitt <nigel.megitt@bbc.co.uk>